removing files that should be git ignored and making colors work. Also dropped AdminModel and Klass and just made Model work
