+++ app/views/upmin/models/dashboard.html.haml
@@ -7,5 +7,3 @@
           This is embarrassing. :(
       %h3
         ^ Click one of those links up top to use features that are working. ^
-      %h5
-        = Upmin::AdminModel.all
