+++ lib/upmin/version.rb
@@ -1,3 +1,3 @@
 module Upmin
-  VERSION = "0.0.39dev10000093"
+  VERSION = "0.0.39dev10000094"
 end
