+++ test_app_upmin/models/admin_user.rb
@@ -1,4 +1,4 @@
-class AdminUser < Upmin::AdminModel
+class AdminUser < Upmin::Model
 
   action :issue_coupon
