--- test_apps/active_record_41/app/upmin/models/admin_shipment.rb
@@ -1,11 +1,0 @@
-class AdminShipment < Upmin::AdminModel
-
-  attribute :status
-  actions :update_shipment
-
-  def status
-    return "TestStatus"
-    # return Upmin::Widget::ProgressBar.new(model.status, model.tracking_states)
-  end
-
-end
