--- test_apps/active_record_42/app/upmin/models/admin_user.rb
@@ -1,5 +1,0 @@
-class AdminUser < Upmin::AdminModel
-
-  action :issue_coupon
-
-end
