--- test_apps/will_paginate/app/upmin/models/admin_product.rb
@@ -1,11 +1,0 @@
-class AdminProduct < Upmin::AdminModel
-  attributes :name, :short_desc, :price, :manufacturer, :free_shipping
-
-  action :update_price
-
-  def update_price(price)
-    model.price = price
-    model.save!
-  end
-
-end
